+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Fixed issue where clones of XFS volumes could not be mounted on the same node as the source volume (Issue [#514](https://github.com/NetApp/trident/issues/514)).
 
 **Enhancements:**
+- **Kubernetes:** Added support for Kubernetes 1.22.
+- **Kubernetes:** Enabled the Trident operator and Helm chart to work with Kubernetes 1.22 (Issue [#628](https://github.com/NetApp/trident/issues/628)).
 
 **Experimental Enhancements:**